Move downloadfile to frontend-core
