+++ packages/builder/src/components/start/ExportAppModal.svelte
@@ -8,6 +8,7 @@
     Input,
     notifications,
   } from "@budibase/bbui"
+  import { downloadFile } from "@budibase/frontend-core"
   import { createValidationStore } from "helpers/validation/yup"
 
   export let app
@@ -55,42 +56,11 @@
   const exportApp = async () => {
     const id = published ? app.prodId : app.devId
     const url = `/api/backups/export?appId=${id}`
-    await downloadFile(url, {
-      excludeRows: !includeInternalTablesRows,
-      encryptPassword: password,
-    })
-  }
-
-  async function downloadFile(url, body) {
     try {
-      const response = await fetch(url, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(body),
+      await downloadFile(url, {
+        excludeRows: !includeInternalTablesRows,
+        encryptPassword: password,
       })
-
-      if (response.ok) {
-        const contentDisposition = response.headers.get("Content-Disposition")
-
-        const matches = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/.exec(
-          contentDisposition
-        )
-
-        const filename = matches[1].replace(/['"]/g, "")
-
-        const url = URL.createObjectURL(await response.blob())
-
-        const link = document.createElement("a")
-        link.href = url
-        link.download = filename
-        link.click()
-
-        URL.revokeObjectURL(url)
-      } else {
-        notifications.error("Error exporting the app.")
-      }
     } catch (error) {
       notifications.error(error.message || "Error downloading the exported app")
     }
